# docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skafolio API",
    "description": "Digital Business Card Platform API - Fast, secure, and feature-rich endpoints for card management, orders, payments, and analytics.",
    "version": "1.0.0",
    "contact": {
      "name": "Skafolio Support",
      "email": "support@skafolio.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "https://skafolio.vercel.app/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/cards/{id}": {
      "get": {
        "tags": ["Cards"],
        "summary": "Get card by ID",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card details"
          }
        }
      },
      "put": {
        "tags": ["Cards"],
        "summary": "Update card",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card updated"
          }
        }
      },
      "delete": {
        "tags": ["Cards"],
        "summary": "Delete card",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card deleted"
          }
        }
      }
    },
    "/cards/{id}/publish": {
      "post": {
        "tags": ["Cards"],
        "summary": "Publish card",
        "description": "Publish card with slug, generate QR code, and create public URL",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["slug"],
                "properties": {
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9-]+$",
                    "example": "john-doe-engineer"
                  },
                  "metaTitle": {
                    "type": "string"
                  },
                  "metaDescription": {
                    "type": "string"
                  },
                  "metaImage": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card published successfully"
          },
          "409": {
            "description": "Slug already taken"
          }
        }
      }
    },
...
    "/payfast/create": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create PayFast payment",
        "description": "Generate PayFast payment URL for an order",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId"],
                "properties": {
                  "orderId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "paymentUrl": {
                      "type": "string"
                    },
                    "paymentId": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/card/{cardId}/summary": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get card analytics summary",
        "description": "Get aggregated analytics for a card. Cached for 5 minutes.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "analytics": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "last24h": {
                          "type": "integer"
                        },
                        "last7d": {
                          "type": "integer"
                        },
                        "last30d": {
                          "type": "integer"
                        },
                        "byType": {
                          "type": "object"
                        }
                      }
                    },
                    "cached": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "CreateCardInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "personal": {
                "type": "object"
              },
              "social": {
                "type": "array"
              },
              "links": {
                "type": "array"
              },
              "stats": {
                "type": "object"
              }
            }
          }
        }
      },
...
    }
  }
}